Refactor de codigo , se crea utils y se mueve modelo dentro de PostsScreenViewModel
